drivers: media: fimc-is-cis-gm2: Fix 'for' loop initial declaration for sensor_gm2_i2c_read16() and sensor_gm2_cis_GGC_write()
Despite the module not being compiled in C99 or C11 mode, this slipped in to the source. It's also the same thing on the M30s drop which makes me more confused. This causes the compiler to fail with "'for' loop initial declarations are only allowed in C99 or C11 mode".

Let's align this to what we follow, just declare the variable before the loop.

Signed-off-by: John Vincent <git@tenseventyseven.cf>