drivers: sensorhub: Add legacy support for pocket mode sensor
Newer vendors for the Samsung A50 has finally changed the pocket mode node name from "pocket_mode_sensor" to follow its siblings that have used "pocket_mode_lite" since their launch.

This change is honestly too little, too late. Having this change merged means that we have to push aside those with older vendors, which is a lot of users.

Instead of removing the older pocket mode node, simply just add a legacy version of it that the kernel can detect and use if pocket_mode_lite is not present.

Potential fix for #8.

Signed-off-by: John Vincent <git@tenseventyseven.cf>