UPSTREAM: zram: Samsung LRU writeback/RAM Plus v3
This is v3 of the Samsung least-recently used (LRU) ZRAM writeback, known as the RAM Plus feature found on selected devices and ROMs on One UI 3, and devices running One UI 4. This specific revision can be found on January 2022 updates of supported devices and the One UI 4 update for older ones.

This includes block bitmapping as well as some sort of 'locking' mechanism on the swap file the service on the system creates on /data/per_boot.

This has been pulled from the Samsung Galaxy S10's One UI 4 kernel, with reference from the Galaxy A52's October 2021 update running One UI 3.1.

Signed-off-by: John Vincent <git@tenseventyseven.cf>